refactor: 핸들러 및 세션 관련 패키지 추가
