+++ tomcat/src/main/java/org/apache/catalina/handler/RegisterHandler.java
@@ -1,4 +1,4 @@
-package org.apache.catalina;
+package org.apache.catalina.handler;
 
 import static org.reflections.Reflections.log;
