+++ tomcat/src/main/java/org/apache/catalina/session/SessionManager.java
@@ -1,7 +1,8 @@
-package org.apache.catalina;
+package org.apache.catalina.session;
 
 import java.util.HashMap;
 import java.util.Map;
+import org.apache.catalina.Manager;
 
 public class SessionManager implements Manager {
